cleanup: organize templates in the same structure that the are scaffolded
