+++ pkg/plugin/v3/scaffolds/internal/templates/config/api/types.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package api
 
 import (
 	"fmt"
