+++ pkg/plugin/v3/scaffolds/internal/templates/config/certmanager/certificate.go
