+++ pkg/plugin/v3/scaffolds/internal/templates/config/certmanager/kustomize.go
