+++ pkg/plugin/v3/scaffolds/internal/templates/config/certmanager/kustomizeconfig.go
