+++ pkg/plugin/v3/scaffolds/internal/templates/config/controller/controller_suitetest.go
