+++ pkg/plugin/v3/scaffolds/internal/templates/config/crd/enablecainjection_patch.go
