+++ pkg/plugin/v3/scaffolds/internal/templates/config/crd/enablewebhook_patch.go
