+++ pkg/plugin/v3/scaffolds/internal/templates/config/crd/kustomization.go
