+++ pkg/plugin/v3/scaffolds/internal/templates/config/crd/kustomizeconfig.go
