+++ pkg/plugin/v3/scaffolds/internal/templates/config/hack/boilerplate.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package hack
 
 import (
 	"fmt"
