+++ pkg/plugin/v3/scaffolds/internal/templates/config/kdefault/auth_proxy_patch.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package metricsauth
+package kdefault
 
 import (
 	"path/filepath"
