+++ pkg/plugin/v3/scaffolds/internal/templates/config/manager/config.go
