+++ pkg/plugin/v3/scaffolds/internal/templates/config/prometheus/kustomize.go
