+++ pkg/plugin/v3/scaffolds/internal/templates/config/prometheus/monitor.go
