+++ pkg/plugin/v3/scaffolds/internal/templates/config/rbac/crd_editor_rbac.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package rbac
 
 import (
 	"path/filepath"
@@ -22,33 +22,48 @@ import (
 	"sigs.k8s.io/kubebuilder/pkg/model/file"
 )
 
-var _ file.Template = &CRDSample{}
+var _ file.Template = &CRDEditorRole{}
 
-// CRDSample scaffolds a manifest for CRD sample.
-type CRDSample struct {
+// CRDEditorRole scaffolds the config/rbac/<kind>_editor_role.yaml
+type CRDEditorRole struct {
 	file.TemplateMixin
 	file.ResourceMixin
 }
 
 // SetTemplateDefaults implements input.Template
-func (f *CRDSample) SetTemplateDefaults() error {
+func (f *CRDEditorRole) SetTemplateDefaults() error {
 	if f.Path == "" {
-		f.Path = filepath.Join("config", "samples", "%[group]_%[version]_%[kind].yaml")
+		f.Path = filepath.Join("config", "rbac", "%[kind]_editor_role.yaml")
 	}
 	f.Path = f.Resource.Replacer().Replace(f.Path)
 
-	f.IfExistsAction = file.Error
-
-	f.TemplateBody = crdSampleTemplate
+	f.TemplateBody = crdRoleEditorTemplate
 
 	return nil
 }
 
-const crdSampleTemplate = `apiVersion: {{ .Resource.Domain }}/{{ .Resource.Version }}
-kind: {{ .Resource.Kind }}
+const crdRoleEditorTemplate = `# permissions for end users to edit {{ .Resource.Plural }}.
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
 metadata:
-  name: {{ lower .Resource.Kind }}-sample
-spec:
-  # Add fields here
-  foo: bar
+  name: {{ lower .Resource.Kind }}-editor-role
+rules:
+- apiGroups:
+  - {{ .Resource.Domain }}
+  resources:
+  - {{ .Resource.Plural }}
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+  - watch
+- apiGroups:
+  - {{ .Resource.Domain }}
+  resources:
+  - {{ .Resource.Plural }}/status
+  verbs:
+  - get
 `
