+++ pkg/plugin/v3/scaffolds/internal/templates/config/rbac/crd_viewer_rbac.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package templates
+package rbac
 
 import (
 	"path/filepath"
@@ -22,43 +22,39 @@ import (
 	"sigs.k8s.io/kubebuilder/pkg/model/file"
 )
 
-var _ file.Template = &CRDEditorRole{}
+var _ file.Template = &CRDViewerRole{}
 
-// CRDEditorRole scaffolds the config/rbac/<kind>_editor_role.yaml
-type CRDEditorRole struct {
+// CRDViewerRole scaffolds the config/rbac/<kind>_viewer_role.yaml
+type CRDViewerRole struct {
 	file.TemplateMixin
 	file.ResourceMixin
 }
 
 // SetTemplateDefaults implements input.Template
-func (f *CRDEditorRole) SetTemplateDefaults() error {
+func (f *CRDViewerRole) SetTemplateDefaults() error {
 	if f.Path == "" {
-		f.Path = filepath.Join("config", "rbac", "%[kind]_editor_role.yaml")
+		f.Path = filepath.Join("config", "rbac", "%[kind]_viewer_role.yaml")
 	}
 	f.Path = f.Resource.Replacer().Replace(f.Path)
 
-	f.TemplateBody = crdRoleEditorTemplate
+	f.TemplateBody = crdRoleViewerTemplate
 
 	return nil
 }
 
-const crdRoleEditorTemplate = `# permissions for end users to edit {{ .Resource.Plural }}.
+const crdRoleViewerTemplate = `# permissions for end users to view {{ .Resource.Plural }}.
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRole
 metadata:
-  name: {{ lower .Resource.Kind }}-editor-role
+  name: {{ lower .Resource.Kind }}-viewer-role
 rules:
 - apiGroups:
   - {{ .Resource.Domain }}
   resources:
   - {{ .Resource.Plural }}
   verbs:
-  - create
-  - delete
   - get
   - list
-  - patch
-  - update
   - watch
 - apiGroups:
   - {{ .Resource.Domain }}
