+++ pkg/plugin/v3/scaffolds/internal/templates/config/rbac/rbac.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package metricsauth
+package rbac
 
 import (
 	"path/filepath"
@@ -22,29 +22,36 @@ import (
 	"sigs.k8s.io/kubebuilder/pkg/model/file"
 )
 
-var _ file.Template = &ClientClusterRole{}
+var _ file.Template = &KustomizeRBAC{}
 
-// ClientClusterRole scaffolds the config/rbac/client_clusterrole.yaml file
-type ClientClusterRole struct {
+// KustomizeRBAC scaffolds the Kustomization file in rbac folder.
+type KustomizeRBAC struct {
 	file.TemplateMixin
 }
 
 // SetTemplateDefaults implements input.Template
-func (f *ClientClusterRole) SetTemplateDefaults() error {
+func (f *KustomizeRBAC) SetTemplateDefaults() error {
 	if f.Path == "" {
-		f.Path = filepath.Join("config", "rbac", "auth_proxy_client_clusterrole.yaml")
+		f.Path = filepath.Join("config", "rbac", "kustomization.yaml")
 	}
 
-	f.TemplateBody = clientClusterRoleTemplate
+	f.TemplateBody = kustomizeRBACTemplate
+
+	f.IfExistsAction = file.Error
 
 	return nil
 }
 
-const clientClusterRoleTemplate = `apiVersion: rbac.authorization.k8s.io/v1beta1
-kind: ClusterRole
-metadata:
-  name: metrics-reader
-rules:
-- nonResourceURLs: ["/metrics"]
-  verbs: ["get"]
+const kustomizeRBACTemplate = `resources:
+- role.yaml
+- role_binding.yaml
+- leader_election_role.yaml
+- leader_election_role_binding.yaml
+# Comment the following 4 lines if you want to disable
+# the auth proxy (https://github.com/brancz/kube-rbac-proxy)
+# which protects your /metrics endpoint.
+- auth_proxy_service.yaml
+- auth_proxy_role.yaml
+- auth_proxy_role_binding.yaml
+- auth_proxy_client_clusterrole.yaml
 `
