+++ pkg/plugin/v3/scaffolds/internal/templates/config/webhook/kustomization.go
