+++ pkg/plugin/v3/scaffolds/webhook.go
@@ -25,7 +25,7 @@ import (
 	"sigs.k8s.io/kubebuilder/pkg/plugin/internal/machinery"
 	"sigs.k8s.io/kubebuilder/pkg/plugin/scaffold"
 	"sigs.k8s.io/kubebuilder/pkg/plugin/v3/scaffolds/internal/templates"
-	"sigs.k8s.io/kubebuilder/pkg/plugin/v3/scaffolds/internal/templates/webhook"
+	"sigs.k8s.io/kubebuilder/pkg/plugin/v3/scaffolds/internal/templates/config/api"
 )
 
 var _ scaffold.Scaffolder = &webhookScaffolder{}
@@ -80,7 +80,7 @@ You need to implement the conversion.Hub and conversion.Convertible interfaces f
 
 	if err := machinery.NewScaffold().Execute(
 		s.newUniverse(),
-		&webhook.Webhook{Defaulting: s.defaulting, Validating: s.validation},
+		&api.Webhook{Defaulting: s.defaulting, Validating: s.validation},
 		&templates.MainUpdater{WireWebhook: true},
 	); err != nil {
 		return err
